fix type and addr

+++ pkg/ccip/chainaccessor/config.go
@@ -137,7 +137,7 @@ func (a *TONAccessor) getOnRampDestChainConfig(ctx context.Context, block *ton.B
 	if err != nil {
 		return ccipocr3.OnRampDestChainConfig{}, err
 	}
-	result, err := a.client.RunGetMethod(ctx, block, addr, "destChainConfig", dest)
+	result, err := a.client.RunGetMethod(ctx, block, addr, "destChainConfig", uint64(dest))
 	if err != nil {
 		return ccipocr3.OnRampDestChainConfig{}, err
 	}
